📝 Update: post 타입 파일 분리
