+++ src/features/post-management/PostContext.tsx
@@ -1,19 +1,5 @@
 import { createContext, PropsWithChildren, useReducer } from "react";
-
-export type PostState = {
-  title: string;
-  code: string;
-  body: string;
-  summary: string;
-};
-
-export type PostAction =
-  | { type: "SET_TITLE"; payload: string }
-  | { type: "SET_CODE"; payload: string }
-  | { type: "SET_BODY"; payload: string }
-  | { type: "SET_SUMMARY"; payload: string }
-  | { type: "SET_ALL"; payload: PostState }
-  | { type: "RESET" };
+import { PostAction, PostContextProps, PostState } from "./type";
 
 const initialState: PostState = {
   title: "",
@@ -40,11 +26,6 @@ const postReducer = (state: PostState, action: PostAction): PostState => {
       throw new Error("Unhandled action type");
   }
 };
-
-type PostContextProps = {
-  state: PostState;
-  dispatch: React.Dispatch<PostAction>;
-} | null;
 
 export const PostContext = createContext<PostContextProps>(null);
